Move tasks.py into io
