+++ mergekit/merge_methods/passthrough.py
@@ -20,8 +20,8 @@ from torch._tensor import Tensor
 
 from mergekit.common import ImmutableMap, ModelReference
 from mergekit.graph import Task
+from mergekit.io.tasks import GatherTensors
 from mergekit.merge_methods.base import ConfigParameterDef, MergeMethod
-from mergekit.tasks import GatherTensors
 
 
 class PassthroughMergeTask(Task[torch.Tensor]):
